Fix runtime error caused by knox dependency, mime
A similar commit was made in onenorth/keystone to fix this issue.

Knox has its own dependency, mime. Knox is no longer maintained and there is an issue with the way it defined its mime dependency in that it did not lock down mime’s version number. Just 2 months ago, mime updated to version 4.0, which no longer supports below Node 14. Our Keystone sites still run on Node 10 so this results in a runtime error (`SyntaxError: Unexpected identifier` on module export syntax) that can’t be worked around.

Keystone Classic fixed this by updating to a fork of knox called knox-s3, which specifies a version of mime but is otherwise the same. This commit applies the same fix.

Links:
Automattic/knox#326
https://github.com/broofa/mime/blob/main/README.md?plain=1#L21
https://github.com/keystonejs/keystone-classic/blob/master/package.json#L63